+++ CONTRIBUTING.md
@@ -42,14 +42,14 @@ Sometimes, updates to [Teaspoons](https://github.com/DataBiosphere/terra-scienti
 
 To test changes in Teaspoons locally against your local CLI, do the following once you've made changes in your local code and are ready to test them:
 
-1. In Teaspoons, run `./gradlew openApiGenerate` to create the autogenerated Python client locally (the files will be generated in `teaspoons-cli/generated` in that repo). You do not need to move these files (and they will not be checked in to the Teaspoons repo).
+1. Follow [the instructions in the Teaspoons repo for creating the autogenerated Python client locally](https://github.com/DataBiosphere/terra-scientific-pipelines-service/blob/main/README.md#testing-the-cli-locally). 
 2. In this repo, edit the [pyproject.toml](pyproject.toml) file as follows: comment out the line that defines the `terra-scientific-pipelines-service-api-client` dependency, and add a line like `teaspoons-client = {path = "YOUR/LOCAL/PATH/TO/terra-scientific-pipelines-service/python-client/generated"}`
  It should look something like this:
  ```
  teaspoons-client = {path = "/Users/awesomedeveloper/workbench/terra-scientific-pipelines-service/python-client/generated"}
 # terra-scientific-pipelines-service-api-client = "^0.1.0"
 ```
-3. Run `poetry lock` and `poetry install`.
+3. Run `poetry update`. (This is equivalent to running `poetry lock` and then `poetry install`.)
 
 Your CLI should now be importing the `teaspoons_client` via the local code rather than the package from PyPi.
 
@@ -60,6 +60,6 @@ poetry install
 ```
 
 To run the CLI against your locally running instance of Teaspoons:
-4. In Teaspoons, run `./gradlew bootRun` and wait for the service to spin up locally.
+4. Follow [the instructions in the Teaspoons repo for running the service locally](https://github.com/DataBiosphere/terra-scientific-pipelines-service/blob/main/README.md#local-development). 
 5. In this repo, edit the `terralab/.terralab-cli-config` file to point the `TEASPOONS_API_URL` to `http://localhost:8080` instead of the live environment deployment of Teaspoons.
 6. That's it! Run your CLI commands and test until you're confident it's all working as you'd like.
