+++ terralab/utils.py
@@ -8,10 +8,7 @@ import typing
 from functools import wraps
 from pydantic import BaseModel
 
-from teaspoons_client import PipelineWithDetails
-from teaspoons_client.exceptions import ApiException
-
-from terralab.logic.pipelines_logic import get_pipeline_info
+from teaspoons_client import ApiException
 
 
 LOGGER = logging.getLogger(__name__)
@@ -60,31 +57,3 @@ def process_json_to_dict(json_data) -> typing.Optional[dict]:
 def is_valid_local_file(local_file_path: str) -> bool:
     """Validate that the provided local file path exists."""
     return True if os.path.exists(local_file_path) else False
-
-
-def validate_pipeline_inputs(pipeline_name: str, inputs_dict: dict):
-    """Check that all required user inputs are present, and that all file inputs are valid."""
-    # retrieve required pipeline inputs; note currently this endpoint only returns required inputs
-    pipeline_info: PipelineWithDetails = get_pipeline_info(pipeline_name)
-
-    input_error_messages = []
-    for (
-        input_info
-    ) in pipeline_info.inputs:  # input_info is PipelineUserProvidedInputDefinition
-        input_name: str = input_info.name
-        LOGGER.debug(f"Validating input {input_name}")
-        if input_name not in inputs_dict:
-            input_error_messages.append(f"Missing required input {input_name}")
-        else:
-            # input is present in inputs_dict; if it's a file, extract the path and validate
-            if input_info.type == "FILE" and (
-                not is_valid_local_file(inputs_dict[input_name])
-            ):
-                input_error_messages.append(
-                    f"Could not find provided file for input {input_name}: {inputs_dict[input_name]}"
-                )
-
-    if input_error_messages:
-        error_string = "\n\t".join(input_error_messages)
-        LOGGER.error(f"Missing or invalid inputs provided:\n\t{error_string}")
-        exit(1)
