+++ tests/test_utils.py
@@ -41,43 +41,3 @@ def test_is_valid_local_file():
 
     # nonexistent file returns False
     assert not (utils.is_valid_local_file("not a file"))
-
-
-def test_validate_pipeline_inputs(capture_logs):
-    test_pipeline_name = "test_pipeline"
-    test_inputs_dict = {"input1": "value1", "input2": "value2"}
-
-    mock_pipeline_info = mock()
-    mock_input1 = mock()
-    mock_input1.name = "input1"
-    mock_input1.type = "STRING"
-    mock_input2 = mock()
-    mock_input2.name = "input2"
-    mock_input2.type = "STRING"
-    mock_pipeline_info.inputs = [mock_input1, mock_input2]
-
-    when(utils).get_pipeline_info(test_pipeline_name).thenReturn(mock_pipeline_info)
-
-    # Should succeed with valid inputs
-    utils.validate_pipeline_inputs(test_pipeline_name, test_inputs_dict)
-
-    # Should fail with missing input
-    with pytest.raises(SystemExit):
-        utils.validate_pipeline_inputs(test_pipeline_name, {"input1": "value1"})
-
-    assert "Missing or invalid inputs provided" in capture_logs.text
-
-    # Should fail with invalid file input
-    mock_file_input = mock()
-    mock_file_input.name = "file_input"
-    mock_file_input.type = "FILE"
-    mock_pipeline_info.inputs = [mock_file_input]
-
-    when(utils).get_pipeline_info(test_pipeline_name).thenReturn(mock_pipeline_info)
-
-    with pytest.raises(SystemExit):
-        utils.validate_pipeline_inputs(
-            test_pipeline_name, {"file_input": "nonexistent_file.txt"}
-        )
-
-    assert "Missing or invalid inputs provided" in capture_logs.text
